Add files via upload
JP.ID is a Multimedia Programming Foundation course project. We were asked to create an application to help JP.ID manage the operators and show the outline of the JP.ID. This project was carried out by 4 people including me. Projects in this course use the Java programming language. I participated in the creation of the Sign Up Page and Product Showcase Pop Up pages